container server: drop duplicated Remove function
